+++ crates/block-producer/src/block_builder/mod.rs
@@ -12,7 +12,7 @@ use miden_objects::{
     note::{Note, NoteExecutionMode, NoteHeader},
     transaction::{OutputNote, TransactionHeader, TransactionId},
 };
-use miden_proving_service_client::proving_service::block_prover::RemoteBlockProver;
+use miden_remote_prover_client::remote_prover::block_prover::RemoteBlockProver;
 use rand::Rng;
 use tokio::time::Duration;
 use tonic::{service::interceptor::InterceptedService, transport::Channel};