+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 
 - Added support for gRPC server side reflection to all components (#949).
 - Compressed faucet background image (#985).
-- Added support for TLS to miden-proving-service-client (#968).
+- Added support for TLS to `miden-remote-prover-client` (#968).
 - Added support for TLS to faucet's connection to node RPC (#976).
 - Replaced integer-based duration args with human-readable duration strings (#998).
 
@@ -47,7 +47,7 @@
 - Introduced Network Transaction Builder (#840).
 - Added way of executing and proving network transactions (#841).
 - [BREAKING] Add HTTP ACCEPT header layer to RPC server to enforce semver requirements against client connections (#844).
-- Add miden-proving-service and miden-proving-service-client crates (#926).
+- Add miden-remote-prover and miden-remote-prover-client crates (#926).
 
 ### Changes