+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Compressed faucet background image (#985).
 - Added support for TLS to `miden-proving-service-client` (#968).
 - Added support for TLS to faucet's connection to node RPC (#976).
+- Replaced integer-based duration args with human-readable duration strings (#998).
 - [BREAKING] Refactor the `miden-proving-service` proxy status service to use gRPC instead of HTTP (#953).
 
 ## v0.9.2 (2025-06-12)